+++ libs/components/code-examples/src/lib/modules/forms/input-box/basic/example.component.spec.ts
@@ -1,7 +1,6 @@
 import { TestbedHarnessEnvironment } from '@angular/cdk/testing/testbed';
-import { TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
-import { SkyAppTestUtility } from '@skyux-sdk/testing';
 import {
   SkyHelpTestingController,
   SkyHelpTestingModule,
@@ -12,20 +11,19 @@ import { FormsInputBoxBasicExampleComponent } from './example.component';
 
 describe('Basic input box example', () => {
   async function setupTest(options: { dataSkyId: string }): Promise<{
+    fixture: ComponentFixture<FormsInputBoxBasicExampleComponent>;
     inputBoxHarness: SkyInputBoxHarness;
     helpController: SkyHelpTestingController;
   }> {
     const fixture = TestBed.createComponent(FormsInputBoxBasicExampleComponent);
     const loader = TestbedHarnessEnvironment.loader(fixture);
     const helpController = TestBed.inject(SkyHelpTestingController);
+
     const inputBoxHarness = await loader.getHarness(
       SkyInputBoxHarness.with({ dataSkyId: options.dataSkyId }),
     );
 
-    fixture.detectChanges();
-    await fixture.whenStable();
-
-    return { inputBoxHarness, helpController };
+    return { fixture, inputBoxHarness, helpController };
   }
 
   beforeEach(() => {
@@ -40,20 +38,25 @@ describe('Basic input box example', () => {
 
   describe('first name field', () => {
     it('should have the expected label text and stacked values', async () => {
-      const { inputBoxHarness } = await setupTest({
+      const { fixture, inputBoxHarness } = await setupTest({
         dataSkyId: 'input-box-first-name',
       });
+
+      fixture.detectChanges();
 
       await expectAsync(inputBoxHarness.getLabelText()).toBeResolvedTo(
         'First name',
       );
+
       await expectAsync(inputBoxHarness.getStacked()).toBeResolvedTo(true);
     });
 
     it('should have the correct help key', async () => {
-      const { inputBoxHarness, helpController } = await setupTest({
+      const { fixture, inputBoxHarness, helpController } = await setupTest({
         dataSkyId: 'input-box-first-name',
       });
+
+      fixture.detectChanges();
 
       await inputBoxHarness.clickHelpInline();
 
@@ -63,19 +66,18 @@ describe('Basic input box example', () => {
 
   describe('last name field', () => {
     it('should have last name required', async () => {
-      const { inputBoxHarness } = await setupTest({
+      const { fixture, inputBoxHarness } = await setupTest({
         dataSkyId: 'input-box-last-name',
       });
 
-      const inputEl = document.querySelector<HTMLInputElement>(
-        'input.last-name-input-box',
+      fixture.detectChanges();
+
+      const inputEl = await inputBoxHarness.querySelector(
+        '.last-name-input-box',
       );
 
-      if (inputEl) {
-        inputEl.value = '';
-        SkyAppTestUtility.fireDomEvent(inputEl, 'input');
-        SkyAppTestUtility.fireDomEvent(inputEl, 'blur');
-      }
+      await inputEl.setInputValue('');
+      await inputEl.blur();
 
       await expectAsync(inputBoxHarness.hasRequiredError()).toBeResolvedTo(
         true,
@@ -85,9 +87,11 @@ describe('Basic input box example', () => {
 
   describe('bio field', () => {
     it('should have a character limit of 250', async () => {
-      const { inputBoxHarness } = await setupTest({
+      const { fixture, inputBoxHarness } = await setupTest({
         dataSkyId: 'input-box-bio',
       });
+
+      fixture.detectChanges();
 
       const characterCounter = await inputBoxHarness.getCharacterCounter();
 
@@ -98,13 +102,13 @@ describe('Basic input box example', () => {
     });
 
     it('should show hint text', async () => {
-      const { inputBoxHarness } = await setupTest({
+      const { fixture, inputBoxHarness } = await setupTest({
         dataSkyId: 'input-box-bio',
       });
 
-      const hintText = await inputBoxHarness.getHintText();
+      fixture.detectChanges();
 
-      expect(hintText).toBe(
+      await expectAsync(inputBoxHarness.getHintText()).toBeResolvedTo(
         `A brief description of the member's background, such as hometown, school, hobbies, etc.`,
       );
     });
@@ -112,17 +116,19 @@ describe('Basic input box example', () => {
 
   describe('email field', () => {
     it('should show a help popover with the expected text', async () => {
-      const { inputBoxHarness } = await setupTest({
+      const { fixture, inputBoxHarness } = await setupTest({
         dataSkyId: 'input-box-email',
       });
 
+      fixture.detectChanges();
+
       await inputBoxHarness.clickHelpInline();
 
-      const helpPopoverTitle = await inputBoxHarness.getHelpPopoverTitle();
-      expect(helpPopoverTitle).toBe('Privacy notice');
+      await expectAsync(inputBoxHarness.getHelpPopoverTitle()).toBeResolvedTo(
+        'Privacy notice',
+      );
 
-      const helpPopoverContent = await inputBoxHarness.getHelpPopoverContent();
-      expect(helpPopoverContent).toBe(
+      await expectAsync(inputBoxHarness.getHelpPopoverContent()).toBeResolvedTo(
         `We do not share this information with any third parties.`,
       );
     });
@@ -130,18 +136,16 @@ describe('Basic input box example', () => {
 
   describe('favorite color field', () => {
     it('should not allow invalid color to be selected', async () => {
-      const { inputBoxHarness } = await setupTest({
+      const { fixture, inputBoxHarness } = await setupTest({
         dataSkyId: 'input-box-favorite-color',
       });
 
-      const selectEl = document.querySelector<HTMLSelectElement>(
-        '.input-box-favorite-color-select',
-      );
+      fixture.detectChanges();
 
-      if (selectEl) {
-        selectEl.value = 'invalid';
-        selectEl.dispatchEvent(new Event('change'));
-      }
+      const selectEl = await inputBoxHarness.querySelector('select');
+
+      // Select the "invalid" option.
+      await selectEl.selectOptions(7);
 
       await expectAsync(
         inputBoxHarness.hasCustomFormError('invalid'),