+++ libs/components/layout/testing/src/public-api.ts
@@ -22,10 +22,14 @@ export { SkyTextExpandHarness } from './modules/text-expand/text-expand-harness'
 export { SkyTextExpandHarnessFilters } from './modules/text-expand/text-expand-harness-filters';
 export { SkyTextExpandModalHarness } from './modules/text-expand/text-expand-modal-harness';
 
-export { SkyToolbarItemHarness } from './modules/toolbar/toolbar-item-harness';
-export { SkyToolbarItemHarnessFilters } from './modules/toolbar/toolbar-item-harness-filters';
+export { SkyTextExpandRepeaterHarness } from './modules/text-expand-repeater/text-expand-repeater-harness';
+export { SkyTextExpandRepeaterHarnessFilters } from './modules/text-expand-repeater/text-expand-repeater-harness-filters';
+export { SkyTextExpandRepeaterItemHarness } from './modules/text-expand-repeater/text-expand-repeater-item-harness';
+
 export { SkyToolbarHarness } from './modules/toolbar/toolbar-harness';
 export { SkyToolbarHarnessFilters } from './modules/toolbar/toolbar-harness-filters';
+export { SkyToolbarItemHarness } from './modules/toolbar/toolbar-item-harness';
+export { SkyToolbarItemHarnessFilters } from './modules/toolbar/toolbar-item-harness-filters';
 export { SkyToolbarSectionHarness } from './modules/toolbar/toolbar-section-harness';
 export { SkyToolbarSectionHarnessFilters } from './modules/toolbar/toolbar-section-harness-filters';
 export { SkyToolbarViewActionsHarness } from './modules/toolbar/toolbar-view-actions-harness';