+++ libs/sdk/cypress-commands/src/lib/ready-command.ts
@@ -1,6 +1,5 @@
 // eslint-disable-next-line @typescript-eslint/no-namespace
 declare namespace Cypress {
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
   interface Chainable<Subject> {
     skyReady(
       appName?: string,